--- a/test/test_ops.py
+++ b/test/test_ops.py
@@ -1088,10 +1088,24 @@ class TestOps(unittest.TestCase):
     helper_test_op([(4,4)], lambda x: x[:, 1:2][0:1])
     helper_test_op([(4,4)], lambda x: x[:, 1:2][:, 0:1])
 
-  @unittest.skip("this test is broken #862")
-  def test_max_inf(self):
-    n = Tensor([1, float("nan")]).max().numpy()
+  @unittest.skipIf(Device.DEFAULT == "WEBGPU", "WEBGPU does not support nan/inf")
+  def test_max_nan(self):
+    n = Tensor([1, float("nan")]).max(ignore_nan=False).numpy()
     assert math.isnan(n.item()), f"{n.item()} is not nan"
+    n = Tensor([float("nan"), 1]).max(ignore_nan=False).numpy()
+    assert math.isnan(n.item()), f"{n.item()} is not nan"
+
+  @unittest.skipIf(Device.DEFAULT == "WEBGPU", "WEBGPU does not support nan/inf")
+  def test_max_float4_nan(self):
+    n = Tensor(np.array([[1.0, 2.0, np.nan, 4.0], [5.0, 6.0, 7.0, 8.0]], dtype=np.float32)).max(ignore_nan=False).numpy()
+    assert math.isnan(n.item()), f"tinygrad max: {n.item()} is not nan"
+
+  @unittest.skipIf(Device.DEFAULT == "WEBGPU", "WEBGPU does not support nan/inf")
+  def test_nan_equality(self):
+    x = Tensor.full((3, 3), float("nan"))
+    result = (x != x).numpy()
+    expected = np.ones((3, 3))
+    assert np.all(result == expected), f"Unexpected result: {result}"
 
   def test_inf_where(self):
     x = Tensor.full((3, 3), float("inf"))